Update Verify.php
Sometimes in the same unit testing, you'll use the same mock in several function, and you will want to reset the counter. I've then added this functionnality.

it can be called like that
ShortifyPunit::verify($sessionMock)->set()->resetCounter();

it's working on exactly the same way that the getChainedMockCounter which retrieve the number of count, this function instead of retrieve the count, it reset it to 0.